% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{counts}
\alias{counts}
\title{Returns the counts matrix of the dataset}
\usage{
counts(x)
}
\arguments{
\item{x}{an object that describes a dataset}
}
\value{
the matrix representing log1p of the data.
}
