% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_functionnal_analysis.R
\name{criteria_vgam}
\alias{criteria_vgam}
\alias{pval_vgam}
\title{Compute vgam pvalues and aic difference criteria}
\usage{
pval_vgam(data, gene, fam = "gaussian", aic = T)
}
\arguments{
\item{data}{a \code{lineageDEDataSet} with results to be plotted.}

\item{gene}{character, a gene of interest.}

\item{fam}{the distribution assumption of the residuals; etiher "binomial" or "gaussian" (default is "gaussian").}

\item{aic}{logical, if the aic criteria is to be computedor not (default is TRUE).}
}
\value{
returns the pvalue and if asked the aic difference for the gene of intereset.
}
\description{
Perform vgam regression with \code{reg_vgam} function and compute p-values based on a likelihood ratio test as well as a AIC based criteria (computation of the difference of AIC between the null model and the alternative model).
}
