% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankings_comparison.R
\name{elbow_curve}
\alias{elbow_curve}
\title{Elbow curve for a ranking}
\usage{
elbow_curve(ranking, xmax = nrow(ranking), xfit = 0.005 * nrow(ranking))
}
\arguments{
\item{ranking}{a \code{rankingDE} object.}

\item{xmax}{integer, at which rank we cut off the curve. (default is length(ranking))}

\item{xfit}{integer, cutoff for the plot of the tangent at the origin (default is 0.005*length(ranking)).}
}
\value{
returns \itemize{\item{a plot of the distances against the ranks of the elements in the ranking.}\item{rank threshold}}
}
\description{
Plot an elbow curve of the distribution of distances from a \code{rankingDE}. It gives an idea of the shape of the distribution and can help select a threshold for the differntially expressed genes.
}
