% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{g.rank}
\alias{g.rank}
\title{Returns the rank and distance of some genes}
\usage{
g.rank(x, g.list)
}
\arguments{
\item{x}{a \code{rankingDE} object.}
}
\value{
the ranks and criteria values for the genes of interest.
}
