% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankings_comparison.R
\name{kendall}
\alias{kendall}
\title{Compute kendall's tau between two rankings}
\usage{
kendall(ranking1, ranking2, nmax = min(dim(ranking1)[1], dim(ranking2)[1]))
}
\arguments{
\item{ranking1}{a first \code{rankingDE} object.}

\item{ranking2}{a second \code{rankingDE} object.}

\item{nmax}{integer, design the last elemtns where to stop the comparison (default is the minimum length of the two rankings).

 @details Kendall's tau is basically the number of permutation between the ranks of the elements of two rankings : tau = ((number of concordant pairs) - (number of discordant pairs))/N}
}
\value{
returns Kendall's tau.
}
\description{
Compares two \code{rankingDE} by computing their kendall's tau from package \code{VGAM}.
}
