% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_approach.R
\name{likelihood_criteria}
\alias{likelihood_criteria}
\title{Compute pvalues and aic difference criteria}
\usage{
likelihood_criteria(data, gene, reg, pval = T, span = 0.75, df = 3,
  fam = "gaussian")
}
\arguments{
\item{data}{a \code{lineageDEDataSet} with results to be plotted.}

\item{gene}{character, a gene of interest.}

\item{reg}{the regression method to use, either "loess" with the gam package or "n.splines" which fits a spline with gam package.}

\item{pval}{logical, if the likelihood criteria is to be computed or not (default is TRUE).}

\item{span}{numeric, for the loess regression, control the ammount of regularization for the loess regression.}

\item{df}{numeric, smoothing parameter for the n.splines regression.}

\item{fam}{character, for the vgam regression, the distribution assumption of the residuals; etiher "binomial" or "gaussian" (default is "gaussian").}
}
\value{
returns the pvalue and if asked the aic difference for the gene of intereset.
}
\description{
Perform a regression with \code{reg_vgam} or \code{reg_loess} function and compute p-values based on a likelihood ratio test as well as a AIC based criteria (computation of the difference of AIC between the null model and the alternative model).
}
