% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_approach.R
\name{likelihood_rank}
\alias{likelihood_rank}
\title{Vgam ranks}
\usage{
likelihood_rank(data, reg = "loess", aic = T, span = 0.5,
  fam = "gaussian")
}
\arguments{
\item{data}{a \code{lineageDEDataSet} with results to be plotted.}

\item{reg}{the regression method to use, either "loess" with the gam package or "vgam" which fits a spline with loess package.}

\item{aic}{logical, if the aic criteria is to be computedor not (default is TRUE).}

\item{span}{numeric, for the loess regression, control the ammount of regularization for the loess regression.}

\item{fam}{character, for the vgam regression, the distribution assumption of the residuals; etiher "binomial" or "gaussian" (default is "gaussian").}
}
\value{
returns \code{rankingDE} objects: one for the likelihood ratio test pvalues and one for the aic difference criteria if aic == T.
}
\description{
compute vgam ranks with likelihood ratio test p-values and aic difference criteria
}
