% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllHelperFunctions.R
\docType{class}
\name{lineageDEDataSet-class}
\alias{lineageDEDataSet-class}
\alias{lineageDEDataSet}
\alias{counts,lineageDEDataSet-method}
\alias{t,lineageDEDataSet-method}
\alias{w,lineageDEDataSet-method}
\title{Class \code{lineageDEDataSet}}
\usage{
\S4method{counts}{lineageDEDataSet}(x)

\S4method{t}{lineageDEDataSet}(x)

\S4method{w}{lineageDEDataSet}(x)
}
\arguments{
\item{x}{a \code{lineageDEDataSet} object.}
}
\description{
The \code{lineageDEDataSet} class holds data relevant for
  performing gene differential expression analysis with the \code{lineageDE} package, primarily a filtered, normalized count matrix of the data and a vecotr of pseudotimes for the cells as well as a vector of the weigths indicating the probability for each cell to belong to one lineage.
  All \code{lineageDE} methods can take an object of the class
  \code{lineageDEDataSet} as input and will output the same.
}
\section{Methods (by generic)}{
\itemize{
\item \code{counts}: returns the matrix of counts.

\item \code{t}: returns the vector of pseudotimes..

\item \code{w}: returns the vector of weights.
}}

