% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_functionnal_analysis.R
\name{loess_rank}
\alias{loess_rank}
\alias{comp_stats}
\title{Vgam ranks}
\usage{
comp_stats(data, gene, fam, aic = T)
}
\arguments{
\item{data}{a \code{lineageDEDataSet} with results to be plotted.}

\item{gene}{character, a gene of interest.}

\item{fam}{the distribution assumption of the residuals; etiher "binomial" or "gaussian" (default is "gaussian").}

\item{aic}{logical, if the aic criteria is to be computedor not (default is TRUE).}
}
\value{
returns \code{rankingDE} objects: one for the likelihood ratio test pvalues and one for the aic difference criteria if aic == T.
}
\description{
compute vgam ranks with likelihood ratio test p-values and aic difference criteria
}
