% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{logCounts}
\alias{logCounts}
\title{Returns the logcounts matrix of the dataset}
\usage{
logCounts(x)
}
\arguments{
\item{x}{an object that describes a dataset}
}
\value{
the matrix representing log1p of the data.
}
