% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{newlineageDEDataSet}
\alias{newlineageDEDataSet}
\title{Initialize an object of class \code{lineageDEDataSet}}
\usage{
newLineageDEDataSet(logCounts, t, w, ...)
}
\arguments{
\item{logCounts}{matrix. An \code{n} by \code{m} numeric matrix or data
frame giving the filtered, normalized RNA logcounts. In rows are the genes and in columns are the cells. It should contain the names of the cells and the genes in \code{colnames(logCounts)} and \code{rownames(logCounts)}.}

\item{t}{matrix. An  \code{m} by 2 numeric matrix
denoting each cell's reconstructed ordering position or pseudotime for the two lineages to compare.}

\item{w}{matrix. An \code{m} by 2 numeric matrix
denoting each cell's probability to belonng to each lineage. The probability should be convex, i.e The sum of the columns shoudl add up to 1.}
}
\description{
Constructs a \code{lineageDEDataSet} object. Additional helper
  methods for manipulating \code{lineageDEDataSet} objects are  also
  described below.
}
