% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{newrankingDE}
\alias{newrankingDE}
\title{Initialize an object of class \code{rankingDE}}
\usage{
newrankingDE(ranking.df, params, ...)
}
\arguments{
\item{ranking.df}{data.frame. An \code{n} by 2 numeric data.frame with the first column "dist" containing the per gene distance computed for the rankinga and the second column "rank" displaying the rank of the gene.}

\item{params}{list. A list giving details about how the ranking has been obtained: which method, what distance has been used, ...}
}
\description{
Constructs a \code{rankingDE} object. Additional helper
methods for manipulating \code{rankingDE} objects are  also
described below.
}
