% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_regression.R
\name{plot_gene}
\alias{plot_gene}
\alias{plot_gene}
\alias{plot_multigenes}
\title{Plot one gene raw data and its regressions}
\usage{
plot_gene(data, gene, regression = T, MD = T, span = 0.75, npred = F,
  sd.show = F, legend.show = F)

plot_multigenes(data, ranking, subset.genes, MD = F, grid.size = NULL)
}
\arguments{
\item{data}{a \code{lineageDEDataSet} with results to be plotted.}

\item{gene}{character, a gene of interest.}

\item{regression}{logical, if the regression is to be computed and plotted or not (defualt is TRUE).}

\item{MD}{logical, if the null model is to be computed and plotted or not (default is TRUE).}

\item{span}{numeric, a span parameter for the loess regression (default is 0.75, same as the \code{loess} function).}

\item{npred}{logical, if the unshared part of the data is to be plotted or not(default is FALSE).}

\item{sd.show}{logical, if the plot of standard deviation is wanted (default is FALSE).}

\item{ranking}{a \code{rankingDE} object, rankings rows of the genes of interest in the ranking dataframe.}

\item{subset.genes}{character vector, the names of the genes of interest.}

\item{grid.size}{2 by 2 vector, for the number of rows and the number of columns that we want for the plot (the size of grid must #' be greater than the number of genes of interest), default is NULL fo a squared grid.}

\item{data}{a \code{lineageDEDataSet} with results to be plotted.}

\item{MD}{logical, if the plot of null model is wanted (default is FALSE).}
}
\value{
returns \itemize{\item{\code{pl},the visualization of the data}
\item{\code{reg}, the regression objects for both lineages and the null model.}}

a visualization of the genes of interest.
}
\description{
Tools for visualizing gene signals for one given gene.

Tools for visualizing gene signals for sveral ranked genes and display their ranking informations.
}
