% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plots.R
\name{plot_multigenes}
\alias{plot_multigenes}
\title{Plot several genes expression patterns}
\usage{
plot_multigenes(data, ranking, subset.genes = NULL, quick.view = "first",
  reg.f = "loess", span = 0.75, s.df = 4, null.model = F,
  grid.size = NULL)
}
\arguments{
\item{data}{a \code{warpDEDataSet} with results to be plotted.}

\item{ranking}{a \code{rankingDE} object, rankings rows of the genes of interest in the ranking dataframe.}

\item{subset.genes}{character vector, let the user specifies the names of the genes of interest (default is NULL).}

\item{quick.view}{character, "first" (default), "last" : specify if we want to see the 8 first or end genes of the ranking.}

\item{reg.f}{a function to perform regression, either "ns" for natural splines, "loess" or "splines" (default is "loess").}

\item{span}{numeric, a smoothing parameter for the regression function (default is 0.75, see \code{gam::lo} for details).}

\item{s.df}{numeric, a smoothing parameter for the nsplines regregression (default is 4, see \code{splines::s} for details about regularization).}

\item{null.model}{logical, if the plot of null model is wanted (default is FALSE).}

\item{grid.size}{2 by 2 vector, for the number of rows and the number of columns that we want for the plot (the size of grid must #' be greater than the number of genes of interest), default is NULL fo a squared grid.}
}
\value{
a visualization of the genes of interest.
}
\description{
Tools for visualizing gene signals for sveral ranked genes and display their ranking informations.
}
