% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankings_comparison.R
\name{plot_rankings}
\alias{plot_rankings}
\title{Compare two rankings by scatter plotting them}
\usage{
plot_rankings(ranking1, ranking2, nmax = min(dim(ranking1)[1],
  dim(ranking2)[1]))
}
\arguments{
\item{ranking1}{a first \code{rankingDE} object.}

\item{ranking2}{a second \code{rankingDE} object.}

\item{nmax}{integer, design the last elemtns where to stop the comparison (default is the minimum length of the two rankings).}
}
\value{
returns a plot comparing two plots rank by rank.
}
\description{
Compares two \code{rankingDE} by plotting the ranks of their elements against each other.
}
