% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ranking.R
\name{quantiles_shared}
\alias{quantiles_shared}
\title{Compare two rankings by counting the shared elements along the rankings.}
\usage{
quantiles_shared(ranking1, ranking2, quantiles = c(0.01, seq(0.05, 1,
  length.out = 20)), nmax = min(dim(ranking.df(ranking1))[1],
  dim(ranking.df(ranking2))[1]))
}
\arguments{
\item{ranking1}{a first \code{rankingDE} object.}

\item{ranking2}{a second \code{rankingDE} object.}

\item{quantiles}{quantiles of the distribution that we wat to compare (default is is 20-quantiles).}

\item{nmax}{integer, design the last elemtns where to stop the comparison (default is the minimum length of the two rankings).}
}
\value{
returns a dataframe displaying the number of common elements between the two ranking in the cumulative quantiles.
}
\description{
Compares two \code{rankingDE} by computing the shared elements in different cumulative quantiles of the rankings.
}
