% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllHelperFunctions.R
\docType{class}
\name{rankingDE-class}
\alias{rankingDE-class}
\alias{rankingDE}
\alias{ranking.df,rankingDE-method}
\alias{params,rankingDE-method}
\title{Class \code{rankingDE}}
\usage{
\S4method{ranking.df}{rankingDE}(x)

\S4method{params}{rankingDE}(x)
}
\arguments{
\item{x}{a \code{rankingDE} object.}
}
\description{
The \code{rankingDE} class holds da ranking format for the \code{lineageDE} package
 It is crafted to be part of the workflow for differentially expressed gen inference and the exploration of the genes ranking.
}
\section{Methods (by generic)}{
\itemize{
\item \code{ranking.df}: returns the ranking dataframe.

\item \code{params}: returns list of parameters used to create the ranking.
}}

