% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankings_comparison.R
\name{rankings_compare}
\alias{rankings_compare}
\title{Compare two rankings with different means}
\usage{
rankings_compare(ranking1, ranking2, nmax = min(dim(ranking.df(ranking1))[1],
  dim(ranking.df(ranking2))[1]))
}
\arguments{
\item{ranking1}{a first \code{rankingDE} object.}

\item{ranking2}{a second \code{rankingDE} object.}
}
\value{
returns the three measures of similarity between the rankings.
}
\description{
Compares two \code{rankingDE} by plotting them against each othe, computing the number of shared elements in the cumulative quantiles of the rankings and computing their kendall's tau.
}
