% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plots.R
\name{reg_gam}
\alias{reg_gam}
\title{Plot one gene raw data and its regressions}
\usage{
reg_gam(data, gene, reg.f = "loess", span = 0.75, s.df = 4,
  regression = T, null.model = T, npred = F, sd.show = F,
  legend.show = F)
}
\arguments{
\item{data}{a \code{warpDEDataSet} with results to be plotted.}

\item{gene}{character, a gene of interest.}

\item{reg.f}{a function to perform regression, either "ns" for natural splines, "loess" or "s" (default is "loess").}

\item{span}{numeric, a smoothing parameter for the regression function (default is 0.75, see \code{gam::lo} for details).}

\item{s.df}{numeric, a smoothing parameter for the nsplines regregression (default is 4, see \code{splines::s} for details about regularization).}

\item{regression}{logical, if the loess regression is to be computed and plotted or not (default is TRUE).}

\item{null.model}{logical, if the null model is to be computed and plotted or not (default is TRUE).}

\item{npred}{logical, if the unshared part of the data is to be plotted or not(default is FALSE).}

\item{legend.show}{logical, if the legend is wanted (default is FALSE).}
}
\value{
returns \itemize{\item{\code{pl},the visualization of the data}
\item{\code{regs}, the regression objects for both lineages and the null model.}}
}
\description{
Tools for visualizing gene signals for one given gene with the gam framework.
}
