% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plots.R
\name{reg_loess}
\alias{reg_loess}
\title{Plot one gene raw data and its loess regressions}
\usage{
reg_loess(data, gene, regression = T, null.model = T, span = 0.5,
  npred = F, sd.show = F, legend.show = F)
}
\arguments{
\item{data}{a \code{lineageDEDataSet} with results to be plotted.}

\item{gene}{character, a gene of interest.}

\item{regression}{logical, if the loess regression is to be computed and plotted or not (defualt is TRUE).}

\item{null.model}{logical, if the null model is to be computed and plotted or not (default is TRUE).}

\item{span}{numeric, a span parameter for the loess regression (default is 0.5).}

\item{npred}{logical, if the unshared part of the data is to be plotted or not(default is FALSE).}

\item{sd.show}{logical, if the plot of standard deviation is wanted (default is FALSE).}
}
\value{
returns \itemize{\item{\code{pl},the visualization of the data}
\item{\code{reg}, the regression objects for both lineages and the null model.}}
}
\description{
Tools for visualizing gene signals for one given gene.
}
