% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plots.R
\name{reg_vgam}
\alias{reg_vgam}
\title{Plot one gene raw data and its vgam regressions}
\usage{
reg_vgam(data, gene, model = "gaussian", regression = T, npred = F,
  null.model = T, legend.show = F)
}
\arguments{
\item{data}{a \code{lineageDEDataSet} with results to be plotted.}

\item{gene}{character, a gene of interest.}

\item{model}{which disrtibution assumption is made on the residuals; either "negbinomial" or  gausian (default is gaussian).}

\item{regression}{logical, if the loess regression is to be computed and plotted or not (default is TRUE).}

\item{npred}{logical, if the unshared part of the data is to be plotted or not(default is FALSE).}

\item{null.model}{logical, if the null model is to be computed and plotted or not (default is TRUE).}

\item{legend.show}{logical, if the legend is to be shown or not (default is FALSE).}
}
\value{
returns \itemize{\item{\code{pl},the visualization of the data}
\item{\code{reg}, the vgam regression objects for both lineages and the null model.}}
}
\description{
Tools for visualizing gene signals for one given gene and compute vgam regression on the count data.
}
