% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{t}
\alias{t}
\title{Returns the pseudotimes}
\usage{
t(x)
}
\arguments{
\item{x}{an object that describes a dataset.}
}
\value{
the matrix of pseudotimes.
}
\description{
Extract the pseudotimes from a \code{SlingshotDataSet}.
}
