% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plots.R
\name{timeDE}
\alias{timeDE}
\title{Differential Expression analysis according with time}
\usage{
timeDE(data, gene, lineage, span = 0.5)
}
\arguments{
\item{data}{a \code{warpDEDataSet} with results to be plotted.}

\item{gene}{character, a gene of interest.}

\item{lineage}{integer, if working with multi lineages data, it sspecifies the lineage of interest).}
}
\value{
returns \itemize{\item{\code{pl},the visualization of the data}
\item{\code{reg}, the regression objects for the lineage of interest and the null model (flat line).}}
}
\description{
Tools for visualizing gene signals in time for one lineage analysis.
}
