% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_functionnal_analysis.R
\name{vgam_rank}
\alias{vgam_rank}
\title{Vgam ranks}
\usage{
vgam_rank(data, fam = "gaussian", aic = T)
}
\arguments{
\item{data}{a \code{lineageDEDataSet} with results to be plotted.}

\item{fam}{the distribution assumption of the residuals; etiher "binomial" or "gaussian" (default is "gaussian").}

\item{aic}{logical, if the aic criteria is to be computedor not (default is TRUE).}

\item{gene}{character, a gene of interest.}
}
\value{
returns \code{rankingDE} objects: one for the likelihood ratio test pvalues and one for the aic difference criteria if aic == T.
}
\description{
compute vgam ranks with likelihood ratio test p-values and aic difference criteria
}
