% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{w}
\alias{w}
\title{Returns the pseudotimes}
\usage{
w(x)
}
\arguments{
\item{x}{an object that describes a dataset.}
}
\value{
the matrix of weights.
}
\description{
Extract the weights (probabilities to beloing to a lineage for a cell) from a \code{SlingshotDataSet}.
}
