% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{warpDE_app}
\alias{warpDE_app}
\title{Interactive plot for a ranking}
\usage{
warpDE_app(data, ranking, nb_genes = 50)
}
\arguments{
\item{data}{a \code{warpDEDataSet} object with genes as rows and cells as columns.}

\item{ranking}{a \code{rankingDE} object , the ranking of the genes of interest that we are interested in.}

\item{nb_genes}{numeric, the number of genes of interest in the list, (default is 50).}
}
\value{
returns nothing.
}
\description{
Given a ranking of genes of a scRNA-Seq experiment, open an interactive interface of the best n genes given the ranking (where n is interactively chosen by the user).
}
