% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warpDE_rank.R
\name{warpDE_rank}
\alias{warpDE_rank}
\title{Wrapper for all ranking method}
\usage{
warpDE_rank(data, reg.f = "loess", span = 0.75, splines.df = 4,
  ranking_method = "warpDE")
}
\arguments{
\item{data}{a \code{warpDEDataSet} with results to be plotted.}

\item{reg.f}{character, the regression method to use, either "loess" with the gam package, "ns" which fits a natural cubic spline with gam package or "splines" which fits a smoothing spline (default is "loess").}

\item{span}{numeric, a smoothing parameter for the regression function if the loess regression (reg.f = 'loess') is used (default is 0.75, see \code{gam::lo} for details about regularization).}

\item{splines.df}{numeric, a smoothing parameter for the nregression function if natural cubic splines (reg.f = 'ns') are used (default is 4, see \code{splines::s} for details about regularization).}

\item{ranking_method}{character, the ranking method to use, either "lrt" for likelihood ratio test, "dtw" for an ordering based solely on dynamic time warping, "warpDE" for a prealable alignment of gene expression with dtw followed by a likelihood ratio test (default is "warpDE").}
}
\value{
returns \code{rankingDE} objects: one for the likelihood ratio test pvalues and one for the aic difference criteria if aic == T.
}
\description{
A wrapper method which takes as input a \code{warpDEDataSet} and various options to return the ranking of the genes following one of three methods available: likelihood ratio test, dynamic time warping distance, or dynamic time warping alignment followed by a likelihood ratio test.
}
