% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllHelperFunctions.R
\docType{class}
\name{warpDEDataSet-class}
\alias{warpDEDataSet-class}
\alias{warpDEDataSet}
\alias{counts,warpDEDataSet-method}
\alias{t,warpDEDataSet-method}
\alias{w,warpDEDataSet-method}
\title{Class \code{warpDEDataSet}}
\usage{
\S4method{counts}{warpDEDataSet}(x)

\S4method{t}{warpDEDataSet}(x)

\S4method{w}{warpDEDataSet}(x)
}
\arguments{
\item{x}{a \code{warpDEDataSet} object.}
}
\description{
The \code{warpDEDataSet} class holds data relevant for
  performing gene differential expression analysis with the \code{warpDE} package, primarily a filtered, normalized count matrix of the data and a vecotr of pseudotimes for the cells as well as a vector of the weigths indicating the probability for each cell to belong to one lineage.
  All \code{warpDE} methods can take an object of the class
  \code{warpDEDataSet} as input and will output the same.
}
\section{Methods (by generic)}{
\itemize{
\item \code{counts}: returns the matrix of counts.

\item \code{t}: returns the vector of pseudotimes..

\item \code{w}: returns the vector of weights.
}}

