% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{warpDEDataSet}
\alias{warpDEDataSet}
\title{warpDEDataSet constructor}
\usage{
warpDEDataSet(counts, t, w)
}
\arguments{
\item{data.frame}{expression matrix}

\item{matrix}{pseudotimes values}

\item{matrix}{weights values}
}
\description{
user friendly constructor for warpDEDataSet
}
