% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtw_ranking.R
\name{dtw_align}
\alias{dtw_align}
\title{Align with dtw two lineages}
\usage{
dtw_align(data, gene, reg.f = "loess", span = 0.75, s.df = 4,
  norm = "L2", window.type = "none", window.size = 50, align.show = F,
  legend.show = F, duplicate.keep = F)
}
\arguments{
\item{data}{a \code{warpDEDataSet} with genes to be ranked.}

\item{gene}{character, a gene of interest.}

\item{reg.f}{a function to perform regression, either "ns" for natural splines, "loess" or "splines" (default is "loess").}

\item{span}{numeric, a smoothing parameter for the regression function (default is 0.75, see \code{gam::lo} for details).}

\item{s.df}{numeric, a smoothing parameter for the nsplines regregression (default is 4, see \code{splines::s} for details about regularization).}

\item{norm}{character,("L2" or "L1") the norm to be used for the dtw distance (default is "L2").}

\item{window.type}{charcater, the window type for the computation of dtw, (see \code{dtw::dtw} for precisions; default is "none").}

\item{window.size}{integer, the size of the warping window (default is NULL, no window.size), see details.}

\item{align.show}{logical, if we want to plot the alignment.}

\item{legend.show}{logical, if the legend is wanted (default is FALSE).}
}
\value{
returns a \itemize{\item \code{pl} a plotof the aligned lineages. \item \code{reg} the alternative and null model for likelihood comaprison.}
}
\description{
Align two lineage with dtw, fit the aligned data and plot the results
}
