% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankings_comparison.R
\name{kendall.heatmap}
\alias{kendall.heatmap}
\title{Heatmap visualization of multiple kendall comparisons}
\usage{
kendall.heatmap(rankings, labels = NULL, addCoef.col = "black",
  cl.pos = "n", ...)
}
\arguments{
\item{labels}{character, a charcater vector for renaming the rankings to compare (default will use the  parameters of the rankings).}

\item{...}{other parameters for the function \code{corrplot}.}

\item{ranking}{list, a list of \code{rankingDE} objects.}
}
\value{
returns a heatmap of kendall's tau
}
\description{
Compute pairwise kendall taus between the elements of a list from a \code{rankingDE}.
}
