% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_approach.R
\name{likelihood_rank}
\alias{likelihood_rank}
\title{Likelihood ranks}
\usage{
likelihood_rank(data, reg.f = "loess", pval = F, span = 0.75, s.df = 4,
  fam = "gaussian", dtw = F)
}
\arguments{
\item{data}{a \code{warpDEDataSet} with results to be plotted.}

\item{reg.f}{the regression method to use, either "loess" with the gam package, "ns" which fits a natural cubic spline with gam package or "splines" which fits a smoothing spline (default is "loess").}

\item{pval}{logical, if the pval criteria is to be computedor not (default is F).}

\item{span}{numeric, a smoothing parameter for the regression function (default is 0.75, see \code{gam::lo} for details).}

\item{s.df}{numeric, a smoothing parameter for the nsplines regregression (default is 4, see \code{splines::s} for details about regularization).}

\item{fam}{character, for the vgam regression, the distribution assumption of the residuals; etiher "binomial" or "gaussian" (default is "gaussian").}

\item{dtw}{logical, if the alignment should be made with dtw before acomputing the likelihood, (default is FALSE) .}
}
\value{
returns \code{rankingDE} objects: one for the likelihood ratio test pvalues and one for the aic difference criteria if aic == T.
}
\description{
compute vgam ranks with likelihood ratio test p-values and aic difference criteria
}
