% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plots.R
\name{plot_multigenes}
\alias{plot_multigenes}
\title{Plot several genes expression patterns}
\usage{
plot_multigenes(data, ranking, subset.genes, reg.f = "ns", span = 0.75,
  df = 4, null.model = F, grid.size = NULL)
}
\arguments{
\item{data}{a \code{lineageDEDataSet} with results to be plotted.}

\item{ranking}{a \code{rankingDE} object, rankings rows of the genes of interest in the ranking dataframe.}

\item{subset.genes}{character vector, the names of the genes of interest.}

\item{reg.f}{a function to perform regression, either "ns" for natural splines, "loess" or "splines" (default is ns).}

\item{span}{numeric, a smoothing parameter for the regression function (default is 0.75, see \code{gam::lo} for details).}

\item{null.model}{logical, if the plot of null model is wanted (default is FALSE).}

\item{grid.size}{2 by 2 vector, for the number of rows and the number of columns that we want for the plot (the size of grid must #' be greater than the number of genes of interest), default is NULL fo a squared grid.}

\item{s.df}{numeric, a smoothing parameter for the nsplines regregression (default is 4, see \code{splines::s} for details about regularization).}
}
\value{
a visualization of the genes of interest.
}
\description{
Tools for visualizing gene signals for sveral ranked genes and display their ranking informations.
}
