% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{warpDE_app}
\alias{warpDE_app}
\title{Interactive plot for a ranking}
\usage{
warpDE_app(data, ranking)
}
\arguments{
\item{data}{a \code{warpDEDataSet} with genes as rows and cells as columns.}

\item{ranking}{a ranking of the gens of interest.}
}
\value{
returns nothing.
}
\description{
Given a ranking of genes of a scRNA-Seq experiment, open an interactive interface of the best n genes given the ranking (where n is interactively chosen by the user).
}
